some fixis
