+++ src/features/subscriptions/ui/BusinessNoSubscriptionView/BusinessNoSubscriptionView.tsx
@@ -1,14 +1,10 @@
 import React from 'react'
 
-import {
-  AccountTypeValue,
-  SubscriptionPlanValue,
-  UISubscriptionPlan,
-} from '@/features/profile/settings/model/types'
 import { AccountTypeSection } from '@/features/profile/settings/ui/AccountManagement/AccountTypeSection/AccountTypeSection'
 
 import styles from './BusinessNoSubscriptionView.module.scss'
 
+import { AccountTypeValue, SubscriptionPlanValue, UISubscriptionPlan } from '../../model/types'
 import { SubscriptionPricing } from '../SubscriptionPricing'
 
 interface BusinessNoSubscriptionViewProps {
