+++ src/features/subscriptions/ui/PersonalView/PersonalView.tsx
@@ -1,10 +1,8 @@
 import React from 'react'
 
-import { AccountTypeValue } from '@/features/profile/settings/model/types'
 import { AccountTypeSection } from '@/features/profile/settings/ui/AccountManagement/AccountTypeSection/AccountTypeSection'
-import { Card, Typography } from '@/shared/ui'
 
-import styles from './PersonalView.module.scss'
+import { AccountTypeValue } from '../../model/types'
 
 interface PersonalViewProps {
   accountType: AccountTypeValue
